requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0

# Kafka
confluent-kafka==2.3.0

# Google Cloud
google-cloud-bigquery==3.14.1
google-auth==2.25.2

# Sentiment Analysis
textblob==0.17.1

# Dashboard
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
db-dtypes==1.1.1
pyarrow==12.0.0

# Utilities
pytz==2023.3